+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes some translations and phrases in the audio settings.
 - [Issue 746](https://github.com/aza547/wow-recorder/issues/746) - Rewords the "Reset Game" button to "Auto-Fit Game". 
 - [Issue 745](https://github.com/aza547/wow-recorder/issues/745) - Fix PTT/Manual hotkeys being overly permissive.
+- [Issue 747](https://github.com/aza547/wow-recorder/issues/747) - Better client logout handling on disabled/deleted guild.
 
 ## [7.0.2] - 2025-10-01
 ### Added